+++ library/Demos/Python/FieldUtils/bfs_vort_rng.py
@@ -2,9 +2,9 @@
 import sys
 from NekPy.FieldUtils import *
 
-field = Field(sys.argv, domain="-1,1,-1,1", force_output=True, error=True)
+field = Field(sys.argv, force_output=True, error=True)
 
-InputModule.Create("xml", field, "bfs_tg.xml").Run()
+InputModule.Create("xml", field, "bfs_tg.xml", range="-1,1,-1,1").Run()
 InputModule.Create("fld", field, "bfs_tg.fld").Run()
 ProcessModule.Create("vorticity", field).Run()
 OutputModule.Create("dat", field, "bfs_tg_vort.fld").Run()