+++ library/Demos/Python/FieldUtils/cube_prismhex_range.py
@@ -2,7 +2,8 @@
 import sys
 from NekPy.FieldUtils import *
 
-field = Field(sys.argv, domain="0,0.5,0,0.5,0,0.5", force_output=True)
+field = Field(sys.argv, force_output=True)
 
-InputModule.Create("xml", field,  "cube_prismhex.xml").Run()
+InputModule.Create("xml", field,  "cube_prismhex.xml",
+                   range="0,0.5,0,0.5,0,0.5").Run()
 OutputModule.Create("xml", field, "out.xml").Run()