+++ CHANGELOG.md
@@ -50,6 +50,9 @@ v5.6.0
 **IncNavierStokesSolver**
 - Matrix-Free LinearADR operator for VCSImplicit and others (!1627)
 
+**FieldConvert**
+- Add range function as an option in Xml Input, align python usage and start depracation of -r option (!1791) 
+
 v5.5.0
 ------
 **Library**